+++ src/app/components/modals/file-chunks/migrate-popover.js
@@ -107,7 +107,9 @@ export default Component.extend(ClickOutside, {
   
   actions: {
     startMigration(destination) {
-      this.get('startMigration')(this.get('sourceProvider.id'), destination);
+      this.get('startMigration')(this.get('sourceProvider.id'), destination, {
+        transfersPending: this.get('migrationTransfersPending'),
+      });
     },
   },
 });